backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# PDF Processing
pypdf==4.0.1

# OpenAI Integration
openai==1.10.0

# Database (opcional para futuro)
sqlalchemy==2.0.25

# Supabase Integration
supabase==2.27.3

# Environment Variables
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# CORS
python-multipart==0.0.6
